requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
httpx>=0.23.0
pinecone>=2.2.4
numpy>=1.24.0
pypdf>=3.10.0
pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest-asyncio>=1.0.0
trio>=0.22.0
python-docx
striprtf
textract
odfpy
pandas
PyYAML
rarfile
